backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
langchain
langchain-community
langchain-groq
Gitpython
sentence-transformers
chromadb
pydantic
httpx
pydantic[email]
